Fixed compiler warnings in tsdf_integrator.h and block.h.

+++ voxblox/include/voxblox/core/block.h
@@ -38,16 +38,16 @@ class Block {
   inline size_t computeLinearIndexFromVoxelIndex(
       const VoxelIndex& index) const {
     size_t linear_index =
-        index.x() +
+        static_cast<size_t>(index.x()) +
         voxels_per_side_ * (index.y() + index.z() * voxels_per_side_);
 
-    DCHECK(index.x() >= 0 && index.x() < voxels_per_side_);
-    DCHECK(index.y() >= 0 && index.y() < voxels_per_side_);
-    DCHECK(index.z() >= 0 && index.z() < voxels_per_side_);
+    DCHECK(index.x() >= 0 && index.x() < static_cast<int>(voxels_per_side_));
+    DCHECK(index.y() >= 0 && index.y() < static_cast<int>(voxels_per_side_));
+    DCHECK(index.z() >= 0 && index.z() < static_cast<int>(voxels_per_side_));
 
     DCHECK_LT(linear_index,
               voxels_per_side_ * voxels_per_side_ * voxels_per_side_);
-    DCHECK_GE(linear_index, 0);
+    DCHECK_GE(linear_index, 0u);
     return linear_index;
   }
